Rename SideBar->Sidebar & rename data_handler->data_handling
