+++ components/search_result.py
@@ -2,7 +2,7 @@ import dash_html_components as html
 import dash_table
 from dash.dependencies import Input, Output
 
-from data_handler import DataFilterer, LoadedData
+from data_handling import DataFilterer, LoadedData
 
 
 # TODO Make callback return a dictionary with inputs
